+++ core/Lucy/Index/SortWriter.c
@@ -93,7 +93,7 @@ S_lazy_init_field_writer(SortWriter *self, int32_t field_num) {
     SortWriterIVARS *const ivars = SortWriter_IVARS(self);
 
     SortFieldWriter *field_writer
-        = (SortFieldWriter*)Vec_Fetch(ivars->field_writers, field_num);
+        = (SortFieldWriter*)Vec_Fetch(ivars->field_writers, (size_t)field_num);
     if (!field_writer) {
 
         // Open temp files.
